fix(desktop): keep compatibility views outside settings monolith
Problem: compatibility account views pushed settings_app.go beyond its repolint file budget.\nFix: extract legacy family view construction into a focused compatibility file and tighten account mutation comments.\nVerification: go run ./tools/repolint; git diff --check; focused desktop settings/model tests.
